frontend: Move OwnershipsList component to ApartmentNewSale directory
- At least so far its only used on that page, and it has references to apartment sales
